requirements.txt
flask
requests
gunicorn
flask-cors
python-dotenv
